enable sidebar for free-tier spaces
