+++ components/common/PageLayout/SharedPageLayout.tsx
@@ -7,25 +7,23 @@ import { useRouter } from 'next/router';
 import useSWR from 'swr';
 
 import charmClient from 'charmClient';
-import { DocumentPageProviders } from 'components/[pageId]/DocumentPage/DocumentPageProviders';
 import { Button } from 'components/common/Button';
-import { PageDialogProvider } from 'components/common/PageDialog/hooks/usePageDialog';
 import { PageDialogGlobal } from 'components/common/PageDialog/PageDialogGlobal';
 import { AppBar } from 'components/common/PageLayout/components/AppBar';
 import CurrentPageFavicon from 'components/common/PageLayout/components/CurrentPageFavicon';
 import { PageTitleWithBreadcrumbs } from 'components/common/PageLayout/components/Header/components/PageTitleWithBreadcrumbs';
 import { HeaderSpacer, StyledToolbar } from 'components/common/PageLayout/components/Header/Header';
 import PageContainer from 'components/common/PageLayout/components/PageContainer';
-import { RewardBlocksProvider } from 'components/rewards/hooks/useRewardBlocks';
-import { RewardsBoardProvider } from 'components/rewards/hooks/useRewardsBoardAndBlocks';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { useIsSpaceMember } from 'hooks/useIsSpaceMember';
 import { useSmallScreen } from 'hooks/useMediaScreens';
-import { ProposalBlocksProvider } from 'hooks/useProposalBlocks';
 import { useSharedPage } from 'hooks/useSharedPage';
 import { useUser } from 'hooks/useUser';
 
+import { LayoutProviders } from './components/LayoutProviders';
 import { LoggedOutButtons } from './components/LoggedOutButtons';
+import { useNavigationSidebar } from './components/Sidebar/hooks/useNavigationSidebar';
+import { NavigationSidebarDrawer } from './components/Sidebar/NavigationSidebarDrawer';
 
 const LayoutContainer = styled.div`
   display: flex;
@@ -36,6 +34,7 @@ type Props = {
   children: React.ReactNode;
   basePageId?: string;
   basePageType?: PageType;
+  enableSidebar: boolean;
 };
 
 // We could update this component in future to support other invite types
@@ -64,75 +63,84 @@ function JoinSpaceWithPublicProposals() {
   );
 }
 
-export function SharedPageLayout({ children, basePageId, basePageType }: Props) {
+export function SharedPageLayout({ children, basePageId, basePageType, enableSidebar }: Props) {
   const isMobile = useSmallScreen();
   const { publicPageType } = useSharedPage();
   const { user } = useUser();
 
+  const { open, enableResize, isResizing, sidebarWidth, handleDrawerOpen, handleDrawerClose } = useNavigationSidebar({
+    enabled: enableSidebar
+  });
+
   return (
-    <ProposalBlocksProvider>
-      <RewardBlocksProvider>
-        <RewardsBoardProvider>
-          <DocumentPageProviders>
-            <Head>
-              <CurrentPageFavicon />
-            </Head>
-            <LayoutContainer data-test='public-page-layout' className='app-content'>
-              <AppBar sidebarWidth={0} position='fixed' open={false}>
-                <StyledToolbar variant='dense'>
-                  <Box
-                    sx={{
-                      display: 'flex',
-                      justifyContent: 'space-between',
-                      alignItems: 'center',
-                      gap: 1,
-                      width: '100%'
-                    }}
-                  >
-                    {!isMobile ? (
-                      <Box display='flex' alignItems='center' gap={6}>
-                        <PageTitleWithBreadcrumbs pageId={basePageId} pageType={basePageType} />
-
-                        {publicPageType === 'proposals' && <JoinSpaceWithPublicProposals />}
-                      </Box>
-                    ) : (
+    <LayoutProviders>
+      <Head>
+        <CurrentPageFavicon />
+      </Head>
+      <LayoutContainer data-test='public-page-layout' className='app-content'>
+        {open !== null && (
+          <>
+            <AppBar open={open} sidebarWidth={sidebarWidth} position='fixed'>
+              <StyledToolbar variant='dense'>
+                <Box
+                  sx={{
+                    display: 'flex',
+                    justifyContent: 'space-between',
+                    alignItems: 'center',
+                    gap: 1,
+                    width: '100%'
+                  }}
+                >
+                  {!isMobile ? (
+                    <Box display='flex' alignItems='center' gap={6}>
                       <PageTitleWithBreadcrumbs pageId={basePageId} pageType={basePageType} />
-                    )}
 
-                    <Box>
-                      {user && (
-                        <Button
-                          endIcon={<ArrowRightIcon />}
-                          variant='text'
-                          color='inherit'
-                          href='/'
-                          external // avoid space domain being added
-                        >
-                          Go to my space
-                        </Button>
-                      )}
-                      {!user && <LoggedOutButtons />}
+                      {publicPageType === 'proposals' && <JoinSpaceWithPublicProposals />}
                     </Box>
-                  </Box>
-                </StyledToolbar>
-                {isMobile && publicPageType === 'proposals' && (
-                  <Box sx={{ width: '100%', ml: 1 }}>
-                    <JoinSpaceWithPublicProposals />
-                  </Box>
-                )}
-              </AppBar>
+                  ) : (
+                    <PageTitleWithBreadcrumbs pageId={basePageId} pageType={basePageType} />
+                  )}
 
-              <PageDialogProvider>
-                <PageContainer>
-                  <HeaderSpacer />
-                  {children}
-                  <PageDialogGlobal />
-                </PageContainer>
-              </PageDialogProvider>
-            </LayoutContainer>
-          </DocumentPageProviders>
-        </RewardsBoardProvider>
-      </RewardBlocksProvider>
-    </ProposalBlocksProvider>
+                  <Box>
+                    {user && (
+                      <Button
+                        endIcon={<ArrowRightIcon />}
+                        variant='text'
+                        color='inherit'
+                        href='/'
+                        external // avoid space domain being added
+                      >
+                        Go to my space
+                      </Button>
+                    )}
+                    {!user && <LoggedOutButtons />}
+                  </Box>
+                </Box>
+              </StyledToolbar>
+              {isMobile && publicPageType === 'proposals' && (
+                <Box sx={{ width: '100%', ml: 1 }}>
+                  <JoinSpaceWithPublicProposals />
+                </Box>
+              )}
+            </AppBar>
+            <NavigationSidebarDrawer
+              enabled={enableSidebar}
+              enableResize={enableResize}
+              enableResizing={!!user}
+              isResizing={isResizing}
+              open={open}
+              width={sidebarWidth}
+              closeSidebar={handleDrawerClose}
+            />
+          </>
+        )}
+
+        <PageContainer>
+          <HeaderSpacer />
+          {children}
+          <PageDialogGlobal />
+        </PageContainer>
+      </LayoutContainer>
+    </LayoutProviders>
   );
 }
